refactor: optimize the unit tests of the `factory` package (#370)
* test case refactor for deployments, sts, clonesets and asts

* refactoring of tests

* refactoring and additional comments

* addition of license header

* refactoring

* refactoring

* lint error

* refactoring

* refactor: reorganized the factory tests


---------
